# Conflicts:
#	makefile
#	miniapps/plasma/CMakeLists.txt
#	miniapps/plasma/makefile
#	miniapps/plasma/plasma.hpp

+++ tests/unit/linalg/test_matrix_square.cpp
@@ -23,7 +23,7 @@ TEST_CASE("FormLinearSystem", "[FormLinearSystem]")
    {
       for (int ne = 1; ne <= 4; ++ne)
       {
-         const int n_elements = std::pow(ne, dim);
+         const int n_elements = static_cast<int>(std::pow(ne, dim));
          CAPTURE(dim, n_elements);
          for (int order = 1; order <= 3; ++order)
          {
@@ -91,7 +91,7 @@ TEST_CASE("ParallelFormLinearSystem", "[Parallel], [ParallelFormLinearSystem]")
    {
       for (int ne = 4; ne <= 5; ++ne)
       {
-         const int n_elements = std::pow(ne, dim);
+         const int n_elements = static_cast<int>(std::pow(ne, dim));
          CAPTURE(dim, n_elements);
          for (int order = 1; order <= 3; ++order)
          {
